Add extras_require.test and pin flake8 < 5.
flake8 5 removes the API being used for configuration.
It's unclear to me as of yet whether it is possible to make this setup
work for both Flake8 <5 and Flake8 5. The ROS infra team has not tackled
this yet.

In the meantime, let's try to get the workflow going again using flake8
4.

Also removes the variable indirection for installs_require since it
isn't used to do anything at the moment.